divide code into parts
